Implementing temperature sampling

+++ app/src/main/java/pl/jatustygne/bus/EventCofeeTemperature.java
@@ -4,7 +4,7 @@ package pl.jatustygne.bus;
  * Created by Dawid Drozd aka Gelldur on 11/18/15.
  */
 public class EventCofeeTemperature {
-	public EventCofeeTemperature(final Integer temperature) {
+	public EventCofeeTemperature(final float temperature) {
 		this.temperature = temperature;
 	}
